Move diagnostics types into a separate helper package
We'll tend to need the diagnostics types and functions in other packages
in the SDK that might be imported by the main tfsdk package, so having
these separated out can avoid import cycles in that case.

The public interface to these is still in the main package. The internal
package the interface only for other packages within this module that need
to return or otherwise work with diagnostics without importing the main
tfsdk package.