Allow resource lifecycle functions to receive tfobj types
Since these types are all superset of cty.Value, we can still extract the
raw value and convert it with gocty if requested, but this gives the
provider implementer convenient access to the most detailed representation
when needed.

All other representations can be derived from these within the provider's
own code if needed, so the automatic conversions just help centralize that
conversion logic in most cases so that providers' own code can be fully
type-safe whenever possible.