Continued importing of the legacy SDK stuff
The "tflegacy" package and some related internal packages will together
become the new homes of all of the types that providers used to import
directly from the hashicorp/terraform repository.

This means that the provider code wouldn't be exactly source compatible
after switching to the new SDK, but that it should be a relatively simple
matter to programmatically rewrite references to these types to talk to
tflegacy instead, allowing a provider codebase to be bulk-converted over
using the legacy shims and then gradually switched over to the new
implementation.