Implement UpgradeResourceState for protocol 5
This was optional in the 0.12.0 alphas this SDK implementation was
originally written/tested against, but it became mandatory before final
release as a way to let the SDK handle automatic normalizations without
a schema version upgrade.

This implementation doesn't actually do any upgrading though, and instead
just assumes what is saved is already in good shape. This would not be
acceptable for a real provider implementation, but is good enough for
ongoing experimentation.