tftest: Allow testing against plan and state contents
When we originally implemented tftest, the terraform-json library wasn't
open source yet and so we couldn't use it here. But now it is, it can
serve as a reasonably lightweight way to decode the JSON output from
Terraform CLI so we can make test assertions against the plan and state
file contents without depending on the main file format parsers in
Terraform CLI itself.

The test case inside the test provider is now more extensive in order to
illustrate how these new APIs might be used.