tflegacy: Just alias things from the real old SDK
When we originally built this the legacy SDK was still embedded in the
main Terraform repository and we didn't want to depend on that from here,
and so we just copied some of the types over instead.

However, the SDK is now in its own separate repository and thus we can now
just use its types directly.

However, the tflegacy package does still provide a nice inventory of the
subset of the legacy functionality this SDK is relying on, so we'll keep
it and instead just make it the home of various aliases and wrappers
around the old SDK such that the main SDK code only depends on tflegacy
directly.