tftest: Allow testing against legacy SDK providers too
The test helper package doesn't actually care very much about which SDK is
being used, since it runs the provider code via a plugin RPC channel
anyway.

Therefore we can have a little helper that provides the same wrapping
behavior for a legacy SDK plugin and then use the rest of the test API
functionality as-is.