tftest: Remove SDK dependencies
Having the tftest package depend on individual SDKs is problematic because
it can cause more than one SDK to be loaded into a process at once, even
though only one should be in use at a time.

tftest was only importing the SDK to offer its helper for initializing a
test, so we'll instead move that into the main SDK package and export the
SDK-agnostic parts it was built from.

This also involved removing the unfinished legacy SDK glue code from the
main tfsdk package, because again otherwise we end up with two SDKs in
the program at once and they both try to take control of the tfplugin5
grpc service. Will need to find some other way to meld in legacy resource
types into a modern provider.